+++ plugins/postcss-nesting/index.js
@@ -12,11 +12,6 @@ function transpileSelectors(fromRule, toRule) {
 	toRule.selectors = selectors;
 }
 
-function cleanNode(node) {
-	if (!('before' in node.raws)) node.raws.before = node.parent.raws.before || '';
-	if (!('after' in node.raws)) node.raws.after = node.parent.raws.after || '';
-}
-
 module.exports = postcss.plugin('postcss-nested', function (opts) {
 	var bubble = ['document', 'media', 'supports'];
 	var name   = 'nest';
@@ -30,13 +25,9 @@ module.exports = postcss.plugin('postcss-nested', function (opts) {
 			var root = rule && rule.parent;
 
 			if (root && rule.type === 'rule') {
-				var newrule = postcss.rule({
-					raws: {
-						before:  atrule.raws.before,
-						between: atrule.raws.between,
-						after:   atrule.raws.after
-					}
-				});
+				var newrule = rule.clone().removeAll();
+
+				newrule.source = atrule.source;
 
 				if (atrule.name === name && atrule.params.indexOf('&') !== -1) {
 					atrule.remove();
@@ -44,8 +35,6 @@ module.exports = postcss.plugin('postcss-nested', function (opts) {
 					newrule.selector = atrule.params;
 
 					newrule.append(atrule.nodes);
-
-					newrule.nodes.forEach(cleanNode);
 
 					transpileSelectors(rule, newrule);
 
@@ -58,8 +47,6 @@ module.exports = postcss.plugin('postcss-nested', function (opts) {
 					newrule.selector = rule.selector;
 
 					newrule.append(atrule.nodes);
-
-					newrule.nodes.forEach(cleanNode);
 
 					atrule.removeAll();