+++ plugins/postcss-nesting/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.5 (2015-10-12)
+
+- Updated: Nested rules source map to the parent rule
+- Updated: PostCSS 5.0.9
+- Updated: Tests and documentation
+- Updated: Project configuration
+
 ## 2.0.4 (2015-09-23)
 
 - Updated: Map source raws